+++ linode/framework_provider.go
@@ -3,7 +3,7 @@ package linode
 import (
 	"context"
 
-	"github.com/linode/terraform-provider-linode/v2/linode/vpc_ips"
+	"github.com/linode/terraform-provider-linode/v2/linode/vpcips"
 
 	"github.com/hashicorp/terraform-plugin-framework/datasource"
 	"github.com/hashicorp/terraform-plugin-framework/provider"
@@ -271,7 +271,7 @@ func (p *FrameworkProvider) DataSources(ctx context.Context) []func() datasource
 		kernels.NewDataSource,
 		vpcsubnet.NewDataSource,
 		vpc.NewDataSource,
-		vpc_ips.NewDataSource,
+		vpcips.NewDataSource,
 		vpcsubnets.NewDataSource,
 		vpcs.NewDataSource,
 		volumes.NewDataSource,
