+++ linode/vpcips/framework_schema_datasource.go
@@ -1,4 +1,4 @@
-package vpc_ips
+package vpcips
 
 import (
 	"github.com/hashicorp/terraform-plugin-framework/datasource/schema"
