+++ linode/vpcips/tmpl/data_basic.gotf
